requirements.txt
alembic==1.16.1
annotated-types==0.6.0
anyio==4.3.0
black==25.1.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.6
jose==1.0.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.29
starlette==0.36.3
swagger_ui_bundle==1.1.0
typing_extensions==4.14.0
uvicorn==0.29.0
WTForms==3.1.2
